refactor: Relocate domain components from shared/ to view directories
