+++ apps/ui/src/components/layout/sidebar/dialogs/trash-dialog.tsx
@@ -9,7 +9,7 @@ import {
   DialogTitle,
 } from '@protolabs-ai/ui/atoms';
 import { Button } from '@protolabs-ai/ui/atoms';
-import { DeleteConfirmDialog } from '@/components/ui/delete-confirm-dialog';
+import { DeleteConfirmDialog } from '@/components/shared/delete-confirm-dialog';
 import { ConfirmDialog } from '@protolabs-ai/ui/molecules';
 import type { TrashedProject } from '@/lib/electron';
