+++ apps/ui/src/components/views/board-view/components/ansi-output.tsx
