+++ apps/ui/src/components/views/board-view/components/git-diff-panel.tsx
