+++ apps/ui/src/components/views/board-view/components/log-viewer.tsx
