+++ apps/ui/src/components/views/board-view/components/task-progress-panel.tsx
