+++ apps/ui/src/components/views/board-view/components/xterm-log-viewer.tsx
