+++ apps/ui/src/components/views/settings-view/components/delete-project-dialog.tsx
@@ -1,5 +1,5 @@
 import { Folder } from 'lucide-react';
-import { DeleteConfirmDialog } from '@/components/ui/delete-confirm-dialog';
+import { DeleteConfirmDialog } from '@/components/shared/delete-confirm-dialog';
 import type { Project } from '@/lib/electron';
 
 interface DeleteProjectDialogProps {
